+++ changelog.md
@@ -1,3 +1,10 @@
 
+# 1.1.0
+Critical changes!
+- Align now works in local coordinates, if you pass options.localCoordinates=true
+
+- AlignLocal - Align working in local coordinates
+
+
 # 1.0.34
 Add `Align.anchorArea` method and bump version to 1.0.34